Добавляет классам приставку component
