+++ src/components/empty-component.js
@@ -1,18 +1,21 @@
-import {createElement} from "../utils";
+import {createElement} from '../utils.js';
 
-const createBoardTemplate = () => {
+const createEmptyTemplate = () => {
   return (
-    `<section class="board container"></section>`
+    `<p class="board__no-tasks">
+      Click «ADD NEW TASK» in menu to create your first task
+    </p>`
   );
 };
 
-export default class Board {
+
+export default class EmptyComponent {
   constructor() {
     this._element = null;
   }
 
   getTemplate() {
-    return createBoardTemplate();
+    return createEmptyTemplate();
   }
 
   getElement() {
