+++ src/components/task-component.js
@@ -83,7 +83,7 @@ const createTaskTemplate = (task) => {
   );
 };
 
-export default class Task {
+export default class TaskComponent {
   constructor(task) {
     this._task = task;
     this._element = null;
