+++ src/components/task-list-component.js
@@ -1,21 +1,19 @@
 import {createElement} from '../utils.js';
 
-const createNoTasksTemplate = () => {
+const createTaskListTemplate = () => {
   return (
-    `<p class="board__no-tasks">
-      Click «ADD NEW TASK» in menu to create your first task
-    </p>`
+    `<div class="board__tasks"></div>`
   );
 };
 
 
-export default class NoTasks {
+export default class TaskListComponent {
   constructor() {
     this._element = null;
   }
 
   getTemplate() {
-    return createNoTasksTemplate();
+    return createTaskListTemplate();
   }
 
   getElement() {
